requirements.txt
llama-cpp-python>=0.2.0
requests>=2.28.0
beautifulsoup4>=4.11.0
